requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic>=1.10.0,<2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
gunicorn==21.2.0
asyncpg==0.29.0